--- lab5/lib/process_image.h
@@ -1,6 +1,0 @@
-#ifndef IMAGE_PROCESSING_H
-#define IMAGE_PROCESSING_H
-
-int process_image(char *input_filename, char *output_filename);
-
-#endif  // IMAGE_PROCESSING_H
